% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{active_hepatitis_df}
\alias{active_hepatitis_df}
\title{Chronic Active Hepatitis Clinical Trial}
\format{
A data frame with 44 observations and 3 variables:
\describe{
  \item{treatment}{Integer vector indicating treatment group: 1 for prednisolone, 0 for control}
  \item{time}{Integer vector representing the time to event or censoring (in days)}
  \item{status}{Integer vector indicating status: 1 for death, 0 for censored}
}
}
\source{
Data taken from the collett package version 0.1.0
}
\usage{
data(active_hepatitis_df)
}
\description{
This dataset, active_hepatitis_df, is a data frame containing information from
a clinical trial of 44 patients with chronic active hepatitis. Patients were randomized
to receive either the drug prednisolone or no treatment (control group).
}
\details{
The dataset name has been kept as 'active_hepatitis_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
