% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{aids_azt_df}
\alias{aids_azt_df}
\title{AIDS Symptoms and AZT Use Data}
\format{
A data frame with 4 observations and 4 variables:
\describe{
  \item{yes}{Numeric vector indicating the number of patients showing AIDS symptoms}
  \item{no}{Numeric vector indicating the number of patients not showing AIDS symptoms}
  \item{azt}{Factor with 2 levels indicating AZT use (\code{yes}, \code{no})}
  \item{race}{Factor with 2 levels indicating patient race (\code{white}, \code{black})}
}
}
\source{
Data taken from the cond package version 1.2-4
}
\usage{
data(aids_azt_df)
}
\description{
This dataset, aids_azt_df, is a data frame containing cross-classified counts of AIDS symptoms
and AZT use by race of the patients, as reported in a 1991 New York Times article.
}
\details{
The dataset name has been kept as 'aids_azt_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
