% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{campy_infections_ts}
\alias{campy_infections_ts}
\title{Campylobacter Infections Time Series}
\format{
A time series object of class \code{ts} with 140 observations, frequency 13,
starting from 1990 to 2000 (end of October).
}
\source{
Data taken from the tscount package version 1.4.3. Original study:
Ferland, R., Latour, A. and Oraichi, D. (2006) Integer-valued GARCH process.
\emph{Journal of Time Series Analysis} 27(6), 923--942.
}
\usage{
data(campy_infections_ts)
}
\description{
This dataset, campy_infections_ts, is a time series object containing the number of cases
of campylobacter infections in the north of the province Quebec (Canada) in four week intervals
from January 1990 to the end of October 2000. It contains 13 observations per year and 140
observations in total.
}
\details{
The dataset name has been kept as 'campy_infections_ts' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix 'ts' indicates that the dataset is a time series object. The original content
has not been modified in any way.
}
