% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{china_dengue_tbl_df}
\alias{china_dengue_tbl_df}
\title{Dengue Cases in Mainland China (2005–2020)}
\format{
A tibble with 16 observations and 5 variables:
\describe{
  \item{year}{Integer year of observation (2005–2020)}
  \item{dengue.cases.indigenous}{Numeric vector of indigenous dengue cases}
  \item{dengue.cases.imported}{Numeric vector of imported dengue cases}
  \item{counties.with.dengue.fever.indigenous}{Numeric vector of counties with reported indigenous dengue fever}
  \item{counties.with.dengue.fever.imported}{Numeric vector of counties with reported imported dengue fever}
}
}
\source{
Data taken from the denguedatahub package version 2.1.1
}
\usage{
data(china_dengue_tbl_df)
}
\description{
This dataset, china_dengue_tbl_df, is a tibble containing annual records of
indigenous and imported dengue cases in mainland China from 2005 to 2020.
}
\details{
The dataset name has been kept as 'china_dengue_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble. The original content has not been modified
in any way.
}
