% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{covid_new_york_df}
\alias{covid_new_york_df}
\title{New York City COVID-19 Data}
\format{
A data frame with 615 observations and 5 variables:
\describe{
  \item{date}{Date of observation}
  \item{borough}{Character vector indicating the borough (e.g., Manhattan, Bronx, etc.)}
  \item{case}{Integer vector representing the number of reported COVID-19 cases}
  \item{hospitalization}{Integer vector representing the number of hospitalizations}
  \item{death}{Integer vector representing the number of deaths}
}
}
\source{
Data taken from the incidental package version 0.1
}
\usage{
data(covid_new_york_df)
}
\description{
This dataset, covid_new_york_df, is a data frame containing daily proportions of COVID-19 cases,
hospitalizations, and deaths by borough in New York City through 2020-06-30.
}
\details{
The dataset name has been kept as 'covid_new_york_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
