% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{covid_severity_df}
\alias{covid_severity_df}
\title{COVID-19 Cardiovascular Severity}
\format{
A data frame with 6 observations and 3 variables:
\describe{
  \item{study}{Character vector with the name or reference of each study}
  \item{beta}{Numeric vector representing the estimated effect size (\eqn{\beta})}
  \item{se}{Numeric vector representing the standard error associated with each estimate}
}
}
\source{
Data taken from the PRP package version 0.1.1
}
\usage{
data(covid_severity_df)
}
\description{
This dataset, \code{covid_severity_df}, is a data frame containing several effect
estimates (\eqn{\beta}) and their standard errors for the impact of cardiovascular
disease on the severe case rate of COVID-19 as reported in the literature.
}
\details{
The dataset name has been kept as \code{covid_severity_df} to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix \code{_df} indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
