% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{ebola_sleone_df}
\alias{ebola_sleone_df}
\title{Ebola Cases in Sierra Leone, Africa}
\format{
A data frame with 110 observations and 2 variables:
\describe{
  \item{Day}{Integer indicating the number of days since May 1, 2014}
  \item{Cases}{Integer representing the cumulative number of Ebola cases reported}
}
}
\source{
Data taken from the MMAC package version 0.1.2
}
\usage{
data(ebola_sleone_df)
}
\description{
This dataset, ebola_sleone_df, is a data frame containing the cumulative number of
Ebola virus disease cases in Sierra Leone, Africa, recorded from May 1, 2014 to
December 16, 2015.
}
\details{
The dataset name has been kept as 'ebola_sleone_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
