% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{ebola_survey_tbl_df}
\alias{ebola_survey_tbl_df}
\title{Survey on Ebola Quarantine}
\format{
A tibble with 1,042 observations and 1 variable:
\describe{
  \item{quarantine}{Factor with two levels indicating whether the respondent supports a mandatory
  21-day quarantine for individuals who have come in contact with an Ebola patient}
}
}
\source{
Data taken from the openintro package version 2.5.0
}
\usage{
data(ebola_survey_tbl_df)
}
\description{
This dataset, ebola_survey_tbl_df, is a tibble containing responses from a poll conducted in New York City
between October 26th and 28th, 2014. The poll was conducted shortly after a doctor who had treated Ebola patients
in Guinea was diagnosed with Ebola in New York City. Participants were asked whether they favored a "mandatory
21-day quarantine for anyone who has come in contact with an Ebola patient". The survey included responses from
1,042 adults residing in New York.
}
\details{
The dataset name has been kept as 'ebola_survey_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble.
The original content has not been modified in any way.
}
