% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{flu_enrich_df}
\alias{flu_enrich_df}
\title{Flu Enrichment Gene Data}
\format{
A data frame with 5719 observations and 3 variables:
\describe{
  \item{nflugenes}{Numeric vector representing gene identifiers with an effect on influenza-virus replication}
  \item{setsize}{Integer vector representing the size of each gene set}
  \item{GO_terms}{Factor vector representing Gene Ontology terms associated with each gene set}
}
}
\source{
Data taken from the rvalues package version 0.7.1
}
\usage{
data(flu_enrich_df)
}
\description{
This dataset, flu_enrich_df, is a data frame containing gene-set enrichment information
for genes that have been identified as having an effect on influenza-virus replication.
}
\details{
The dataset name has been kept as 'flu_enrich_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
