% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{hepatitisA_df}
\alias{hepatitisA_df}
\title{Hepatitis A Prevalence in Bulgaria}
\format{
A data frame with 83 observations and 3 variables:
\describe{
  \item{t}{Integer vector indicating the age of the individuals}
  \item{freq1}{Integer vector representing the frequency of individuals tested}
  \item{freq2}{Integer vector representing the frequency of individuals with antibodies to hepatitis A}
}
}
\source{
Data taken from the curstatCI package version 0.1.1
}
\usage{
data(hepatitisA_df)
}
\description{
This dataset, hepatitisA_df, is a data frame containing information from a cross-sectional
survey conducted in 1964 on the prevalence of hepatitis A in individuals from Bulgaria.
The surveyed population includes individuals aged between 1 and 86 years.
}
\details{
The dataset name has been kept as 'hepatitisA_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
