% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{india_dengue_tbl_df}
\alias{india_dengue_tbl_df}
\title{Dengue/DHF Situation in India Since 2017}
\format{
A tibble with 432 observations and 5 variables:
\describe{
  \item{area}{Character vector indicating the State or Union Territory}
  \item{type}{Character vector indicating whether the entry refers to 'cases' or 'deaths'}
  \item{year}{Character vector indicating the year of observation}
  \item{additional_information}{Character vector providing supplemental information}
  \item{value}{Numeric vector indicating the number of cases or deaths}
}
}
\source{
Data taken from the denguedatahub package version 2.1.1
}
\usage{
data(india_dengue_tbl_df)
}
\description{
This dataset, india_dengue_tbl_df, is a tibble containing state and union territory-wise
annual dengue/DHF (Dengue Hemorrhagic Fever) cases and deaths in India since 2017.
}
\details{
The dataset name has been kept as 'india_dengue_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble (enhanced data frame). The original content has not been modified
in any way.
}
