% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infectiousR-package.R
\docType{package}
\name{infectiousR}
\alias{infectiousR}
\alias{infectiousR-package}
\title{infectiousR: Access Infectious and Epidemiological Data via 'disease.sh API'}
\description{
{
This package provides functions to access real-time infectious disease data from the 'disease.sh API', including COVID-19 global, US states, continent, and country statistics, vaccination coverage,influenza-like illness data from Centers for Disease Control and Prevention (CDC), also includes curated datasets on a variety of infectious diseases such as influenza, measles, dengue, Ebola, tuberculosis, meningitis, AIDS, and others.
}
}
\details{
infectiousR: Access Infectious and Epidemiological Data via 'disease.sh API'

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access Infectious and Epidemiological Data via 'disease.sh API'.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/infectiousr}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
