% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{influenza_ice_df}
\alias{influenza_ice_df}
\title{Monthly Influenza Incidence in Iceland}
\format{
A data frame with 360 observations and 3 variables:
\describe{
  \item{month}{Integer vector representing the month of observation (1--12)}
  \item{year}{Integer vector representing the year of observation (1980--2009)}
  \item{ili}{Integer vector representing the monthly incidence of influenza-like illness}
}
}
\source{
Data taken from the epimdr package version 0.6-5
}
\usage{
data(influenza_ice_df)
}
\description{
This dataset, influenza_ice_df, is a data frame containing monthly incidence
data of influenza-like illness (ILI) in Iceland between 1980 and 2009.
}
\details{
The dataset name has been kept as 'influenza_ice_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
