% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{influenza_pneumonia_ts}
\alias{influenza_pneumonia_ts}
\title{US Pneumonia and Influenza Death Rates}
\format{
A time series object with 132 monthly observations:
\describe{
  \item{value}{Monthly pneumonia and influenza deaths per 10,000 people in the United States from 1968 to 1978.}
}
}
\source{
Data taken from the \pkg{astsa} package version 2.2.
}
\usage{
data(influenza_pneumonia_ts)
}
\description{
This dataset, \code{influenza_pneumonia_ts}, is a time series containing monthly pneumonia and influenza deaths
per 10,000 people in the United States over a period of 11 years, from 1968 to 1978.
}
\details{
The dataset name has been kept as \code{influenza_pneumonia_ts} to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
\pkg{infectiousR} package and assists users in identifying its specific characteristics.
The suffix \code{_ts} indicates that the dataset is a time series object. The original content has not been modified
in any way.
}
