% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{rubella_austria_df}
\alias{rubella_austria_df}
\title{Rubella Prevalence in Austrian Males}
\format{
A data frame with 225 observations and 3 variables:
\describe{
  \item{t}{Numeric vector representing age or time (in months or years as recorded)}
  \item{freq1}{Integer vector representing frequency count 1}
  \item{freq2}{Integer vector representing frequency count 2}
}
}
\source{
Data taken from the curstatCI package version 0.1.1
}
\usage{
data(rubella_austria_df)
}
\description{
This dataset, rubella_austria_df, is a data frame containing prevalence data of rubella
in 230 Austrian males older than three months, for whom the exact date of birth was known.
Each individual was tested at the Institute of Virology, Vienna during the period 1–25 March 1988
for immunization against Rubella.
}
\details{
The dataset name has been kept as 'rubella_austria_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
