% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{rubella_peru_df}
\alias{rubella_peru_df}
\title{Rubella in Peru Data}
\format{
A data frame with 95 observations and 4 variables:
\describe{
  \item{age}{Numeric vector indicating the age of individuals}
  \item{incidence}{Integer vector indicating the number of rubella cases per age group}
  \item{cumulative}{Integer vector indicating the cumulative number of cases by age}
  \item{n}{Integer vector representing the sample size for each age group}
}
}
\source{
Data taken from the epimdr package version 0.6-5
}
\usage{
data(rubella_peru_df)
}
\description{
This dataset, \code{rubella_peru_df}, is a data frame containing rubella incidence data
by age as studied by Metcalf et al. (2011) in Peru.
}
\details{
The dataset name has been kept as \code{rubella_peru_df} to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix \code{_df} indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
