% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{sars_canada_df}
\alias{sars_canada_df}
\title{Severe Acute Respiratory Syndrome in Canada, 2003}
\format{
A data frame with 110 observations and 5 variables:
\describe{
  \item{date}{Date object representing the reporting date}
  \item{cases_travel}{Integer vector indicating new SARS cases linked to travel}
  \item{cases_household}{Integer vector indicating new SARS cases from household transmission}
  \item{cases_healthcare}{Integer vector indicating new SARS cases from healthcare settings}
  \item{cases_other}{Integer vector indicating new SARS cases from other or unknown sources}
}
}
\source{
Data taken from the outbreaks package version 1.9.0
}
\usage{
data(sars_canada_df)
}
\description{
This dataset, sars_canada_df, is a data frame containing information on the daily incidence of
SARS (Severe Acute Respiratory Syndrome) cases in Canada during the 2003 outbreak. The data include
new cases attributed to travel, household transmission, healthcare settings, and other sources.
}
\details{
The dataset name has been kept as 'sars_canada_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
