% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{spanish_flu_df}
\alias{spanish_flu_df}
\title{Daily 1918 Flu Deaths}
\format{
A data frame with 122 observations and 4 variables:
\describe{
  \item{Date}{Date of recorded mortality}
  \item{Indiana}{Integer vector representing daily flu-related deaths in Indiana}
  \item{Kansas}{Integer vector representing daily flu-related deaths in Kansas}
  \item{Philadelphia}{Integer vector representing daily flu-related deaths in Philadelphia}
}
}
\source{
Data taken from the incidental package version 0.1
}
\usage{
data(spanish_flu_df)
}
\description{
This dataset, spanish_flu_df, is a data frame containing daily mortality data
from the 1918 flu pandemic covering the period from 1918-09-01 through 1918-12-31
in Indiana, Kansas, and Philadelphia.
}
\details{
The dataset name has been kept as 'spanish_flu_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
