% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{us_covid_cases_df}
\alias{us_covid_cases_df}
\title{US Lab-Confirmed COVID-19 Cases}
\format{
A data frame with 1227 observations and 2 variables:
\describe{
  \item{date}{Date of report (class \code{Date})}
  \item{cases}{Integer vector indicating the number of confirmed cases reported on each date}
}
}
\source{
Data taken from the cpr package version 0.4.0
}
\usage{
data(us_covid_cases_df)
}
\description{
This dataset, \code{us_covid_cases_df}, is a data frame containing the number of
laboratory-confirmed COVID-19 cases in the United States, as reported by the
Centers for Disease Control and Prevention (CDC), between January 1, 2020 and
May 11, 2023, the end of the public health emergency declaration.
}
\details{
The dataset name has been kept as \code{us_covid_cases_df} to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix \code{_df} indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
