% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{zika_sanandres_df}
\alias{zika_sanandres_df}
\title{Zika in San Andres, Colombia, 2015}
\format{
A data frame with 101 observations and 2 variables:
\describe{
  \item{date}{Date object representing the date of reported Zika cases}
  \item{cases}{Integer vector indicating the number of daily reported Zika cases}
}
}
\source{
Data taken from the outbreaks package version 1.9.0
}
\usage{
data(zika_sanandres_df)
}
\description{
This dataset, zika_sanandres_df, is a data frame containing the daily incidence of
Zika virus disease in San Andres, Colombia, during 2015.
}
\details{
The dataset name has been kept as 'zika_sanandres_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
infectiousR package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
